Move application-related classes to appropriate package
