+++ backend-application/src/main/java/eu/socialedge/hermes/backend/serialization/QuantitySerializer.java
@@ -1,4 +1,4 @@
-package eu.socialedge.hermes.backend.core;
+package eu.socialedge.hermes.backend.serialization;
 
 import com.fasterxml.jackson.core.JsonGenerator;
 import com.fasterxml.jackson.databind.SerializerProvider;
