+++ backend-application/src/test/java/eu/socialedge/hermes/backend/serialization/QuantityDeserializerTest.java
@@ -1,4 +1,4 @@
-package eu.socialedge.hermes.backend.core;
+package eu.socialedge.hermes.backend.serialization;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import lombok.val;
